Struct commands::parser::completion::Completion

pub struct Completion<'c, 'n, 't> {
    pub help_symbol: &'n str,
    pub help_text: &'n str,
    pub token: Option<&'t Token<'t>>,
    pub exhaustive: bool,
    pub options: Vec<CompletionOption<'c>>,
}

Represents the result of completing a node.

This may be hinted by a pre-existing token.

If a completion is exhaustive, then only the CompletionOptions provided are valid.

The lifetime parameter 'c refers to the lifetime of this Completion. The lifetime parameter 't refers to the lifetime of the body of text which generated the Token. The lifetime parameter 'n refers to the lifetime of the node which generated the completion.

Fields

help_symbol

Value placeholder for help.

help_text

Main help text.

token

Token used to hint the completion, if provided.

exhaustive

Was this completion exhaustive? If yes, then only the given completion options are valid.

options

The actual completion options.

Methods

impl<'c, 'n, 't> Completion<'c, 'n, 't>

fn new(node: &'n Node<'n>, token: Option<&'t Token<'t>>, exhaustive: bool, complete_options: Vec<&str>, other_options: Vec<&str>) -> Completion<'c, 'n, 't>