Struct commands::parser::completion::CompletionOption

pub struct CompletionOption<'c> {
    pub option_string: &'c str,
    pub complete: bool,
}

Represents a single option returned by complete.

An option may be complete, which means that it represents a syntactically complete parameter value which can be used as-is, whereas incomplete options are not valid values.

The lifetime parameter 'c refers to the lifetime of the owning Completion.

Fields

option_string

String for this option.

complete

True if this option is complete and a valid value.

Methods

impl<'c> CompletionOption<'c>

fn new(option_string: &'c str, complete: bool) -> CompletionOption

CompletionOption constructor