Module commands::parser::nodes

Parser Nodes

Parser nodes represent the grammar that is defined by the currently permissible set of commands and their parameters.

Structs

CommandNode

CommandNodeFields

FlagParameterNode

NamedParameterNode

NodeFields
A parse tree node.

ParameterNameNode

ParameterNameNodeFields

ParameterNodeFields

RepeatableNodeFields

RootNode

SimpleParameterNode

WrapperNode

WrapperNodeFields

Constants

PRIORITY_DEFAULT
The default priority.

PRIORITY_MINIMUM
Minimum priority.

PRIORITY_PARAMETER
The default priority for a parameter.

Traits

Node

ParameterNode