Struct commands::parser::TreeNode

pub struct TreeNode {
    pub name: String,
    pub help_symbol: String,
    pub help_text: String,
    pub hidden: bool,
    pub priority: i32,
    pub repeatable: bool,
    pub repeat_marker: Option<Rc<Node>>,
    pub successors: Vec<Rc<Node>>,
}

A parse tree node.

Fields

name

The name of this node.

help_symbol

The text used to identify this node in help text. This is typically the node name, either in plain form or decorated for parameters.

help_text

Help text describing this node.

hidden

Hidden nodes are not completed. This doesn't modify matching.

priority

Match and complete priority.

repeatable

Whether or not this node can be repeated. A repeated node can be accepted multiple times.

repeat_marker

If present, this node will no longer be acceptable.

successors

Possible successor nodes. Collected while building.