Module commands::parser::tokenizer

Command Tokenization

The command parser needs to be able to tokenize commands into their constituent words and whitespace.

The tokenizer breaks source text into a vector of tokens which can be either whitespace or a word. The tokenizer handles using double quotes to provide a single token which may include whitespace.

Examples

use commands::parser::tokenizer::tokenize;

let tokens = tokenize("word");
assert_eq!(tokens.len(), 1);

let tokens = tokenize("show interface");
assert_eq!(tokens.len(), 3);

let tokens = tokenize("echo -n \"a b c\"");
assert_eq!(tokens.len(), 5);

Structs

SourceLocation
A range within a string.

SourceOffset
A position within a string.

Token
A token from a body of text.

Enums

TokenType
The role that a token plays: Whitespace or Word.

Functions

tokenize
Tokenize a body of text.