Struct commands::parser::Parser

pub struct Parser<'r, 'p> {
    pub nodes: Vec<&'r Node<'r>>,
    pub tokens: Vec<&'p Token<'p>>,
    // some fields omitted
}

Command parser

The lifetime parameter 'r refers to the lifetime of the root node of the command tree that was used to create this parser.

The lifetime parameter 'p refers to the lifetime of the parser and of the tokens passed into this parser.

Fields

nodes

The nodes which have been accepted during parse or advance.

tokens

The tokens which have been accepted during parse or advance.

Methods

impl<'r, 'p> Parser<'r, 'p>

fn new(initial_node: &'r RootNode<'r>) -> Parser<'r, 'p>

Construct a parser with a root node.

fn push_command(&mut self, command: &'r Node<'r>)

XXX: Temporarily public.

fn parse(&mut self, tokens: Vec<&'p Token<'p>>)

Parse a vector of tokens, advancing through the node hierarchy.

fn advance(&mut self, token: &'p Token<'p>)

Parse a single token, advancing through the node hierarchy.